03_intermediario/requirements.txt
asgiref==3.5.2
Django==3.2.15
djangorestframework==3.13.1
Faker==13.15.1
python-dateutil==2.8.2
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
validate-docbr==1.10.0
